+++ src/common/config/index.ts
@@ -1,6 +1,6 @@
 import { CesiumGeographicTilingScheme, Proj } from '@map-colonies/react-components';
 import { IRasterLayer } from '@map-colonies/react-components/dist/cesium-map/layers-manager';
-import { IBaseMaps, IBaseMap } from '@map-colonies/react-components/dist/cesium-map/settings/settings';
+import { IBaseMaps, IBaseMap } from '@map-colonies/react-components/dist/cesium-map/map';
 import { LinkType } from '../models/link-type.enum';
 
 /*eslint-disable */
@@ -40,6 +40,7 @@ const POLYGON_PARTS = {
   billBoardStrokeColor: '#FFFF00'
 };
 const WFS = (window as any)._env_.WFS;
+const GEOCODING_URL = (window as any)._env_.GEOCODING_URL;
 
 const enrichBaseMaps = (baseMaps: IBaseMaps): IBaseMaps => {
   return {
@@ -47,7 +48,7 @@ const enrichBaseMaps = (baseMaps: IBaseMaps): IBaseMaps => {
       return {
         ...baseMap,
         thumbnail: baseMap.thumbnail && ACCESS_TOKEN.injectionType?.toLowerCase() === 'queryparam' ? `${baseMap.thumbnail}?${ACCESS_TOKEN.attributeName}=${ACCESS_TOKEN.tokenValue}` : baseMap.thumbnail,
-        baseRasteLayers: (baseMap.baseRasteLayers as IRasterLayer[]).map((rasterLayer) => {
+        baseRasterLayers: (baseMap.baseRasterLayers as IRasterLayer[]).map((rasterLayer) => {
           return {
             ...rasterLayer,
             options: {
@@ -70,31 +71,31 @@ const systemJobsPriorityOptions =
       label: 'system-status.job-priority-highest',
       value: '2000',
       icon: 'Priority-Highest-Monochrome',
-      iconColor: 'var(--mdc-theme-gc-priority-highest)',
+      iconcolor: 'var(--mdc-theme-gc-priority-highest)',
     },
     {
       label: 'system-status.job-priority-high',
       value: '1500',
       icon: 'Priority-High-Monochrome',
-      iconColor: 'var(--mdc-theme-gc-priority-high)',
+      iconcolor: 'var(--mdc-theme-gc-priority-high)',
     },
     {
       label: 'system-status.job-priority-normal',
       value: '1000',
       icon: 'Priority-Normal-Monochrome',
-      iconColor: 'var(--mdc-theme-gc-priority-normal)',
+      iconcolor: 'var(--mdc-theme-gc-priority-normal)',
     },
     {
       label: 'system-status.job-priority-low',
       value: '500',
       icon: 'Priority-Low-Monochrome',
-      iconColor: 'var(--mdc-theme-gc-priority-low)',
+      iconcolor: 'var(--mdc-theme-gc-priority-low)',
     },
     {
       label: 'system-status.job-priority-lowest',
       value: '0',
       icon: 'Priority-Lowest-Monochrome',
-      iconColor: 'var(--mdc-theme-gc-priority-lowest)',
+      iconcolor: 'var(--mdc-theme-gc-priority-lowest)',
     },
   ];
 
@@ -131,7 +132,9 @@ const APP_CONFIG = {
     ZOOM: MAP.zoom as number,
     MAPMODE2D: MAP.mapMode2D,
     USE_OPTIMIZED_TILE_REQUESTS: MAP.useOptimizedTileRequests as boolean,
-    DEBUG_PANEL: JSON.parse(MAP.debugPanel)
+    SHOW_DEBUGGER_TOOL: MAP.showDebuggerTool,
+    SHOW_ACTIVE_LAYERS_TOOL: MAP.showActiveLayersTool,
+    SHOW_GEOCODER_TOOL: MAP.showGeocoderTool
   },
   ACTIVE_LAYER: ACTIVE_LAYER, // | 'WMTS_LAYER' | 'WMS_LAYER' | 'XYZ_LAYER' | 'OSM_LAYER'
   ACTIVE_LAYER_PROPERTIES: ACTIVE_LAYER_PROPERTIES,
@@ -197,7 +200,7 @@ const APP_CONFIG = {
       MAX_ACTIVE_LAYERS_TO_PRESENT: 5,
       POLYGON_PARTS_FEATURE_CONFIG: {
         color: `${POLYGON_PARTS.highResolutionColor}50`, //'#00ff0030',//'#BF40BF',
-        outlineColor: POLYGON_PARTS.highResolutionColor,//'#00FF00',
+        outlineColor: POLYGON_PARTS.highResolutionColor, //'#00FF00',
         outlineWidth: 8
       }
     }
@@ -232,7 +235,8 @@ const APP_CONFIG = {
       ZOOM_LEVEL: WFS.max.zoomLevel,
       CACHE_SIZE: WFS.max.cacheSize,
     }
-  }
+  },
+  GEOCODING_URL: GEOCODING_URL
 };
 
 export default APP_CONFIG;