+++ src/discrete-layer/components/system-status/cell-renderer/job-details.status.filter.tsx
@@ -1,37 +1,16 @@
-import React, {
-  forwardRef,
-  useImperativeHandle,
-  useState,
-  useEffect,
-  useCallback,
-  useMemo,
-} from 'react';
+import React, { useCallback, useMemo } from 'react';
+import { useIntl } from 'react-intl';
 import { Box } from '@map-colonies/react-components';
 import { Select } from '@map-colonies/react-core';
-import { useIntl } from 'react-intl';
-import { JobModelType, Status } from '../../../models';
-import { IDoesFilterPassParams, IFilterParams, } from 'ag-grid-community';
+import { Status } from '../../../models';
 
-export const JobDetailsStatusFilter = forwardRef((props: IFilterParams, ref) => {
+interface IFilterOnModelChange {
+  onModelChange: (model: any | null, additionalEventAttributes?: any) => void;
+}
+
+export const JobDetailsStatusFilter: React.FC<IFilterOnModelChange> = ({ onModelChange }) => {
   const intl = useIntl();
 
-  const [filterStatus, setFilterStatus] = useState<string | Status>('');
-
-  useEffect(() => {
-    props.filterChangedCallback();
-  }, [props,filterStatus]);
-
-  useImperativeHandle(ref, () => {
-    return {
-      doesFilterPass(params: IDoesFilterPassParams): boolean {
-        return (params.data as JobModelType).status === filterStatus;
-      },
-
-      isFilterActive(): boolean {
-        return filterStatus !== '';
-      },
-    };
-  });
 
   const getStatusTranslation = useCallback((status: Status): string => {
     const statusText = intl.formatMessage({
@@ -39,16 +18,14 @@ export const JobDetailsStatusFilter = forwardRef((props: IFilterParams, ref) => 
     });
 
     return statusText;
-  },[intl]);
+  }, [intl]);
 
   const getStatusOptions = useMemo((): JSX.Element => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const statuses: Record<string, any> = {};
+    const statuses: Record<string, string> = {};
 
     const showAllStatusesText = intl.formatMessage({
       id: 'system-status.job.filter.status.all',
     });
-
 
     for (const [key, val] of Object.entries(Status)) {
       statuses[key] = getStatusTranslation(val);
@@ -56,20 +33,20 @@ export const JobDetailsStatusFilter = forwardRef((props: IFilterParams, ref) => 
 
     return (
       <Select
-        style={{ height: '180px', borderRadius: 0 }}
+        style={{ height: '180px', borderRadius: 0, width: '100%' }}
         enhanced
         placeholder={showAllStatusesText}
         options={statuses}
         onChange={(evt: React.ChangeEvent<HTMLSelectElement>): void => {
-          setFilterStatus(evt.currentTarget.value);
+          onModelChange(evt.currentTarget.value === '' ? null : evt.currentTarget.value)
         }}
       />
     );
-  },[getStatusTranslation, intl]);
+  }, [getStatusTranslation, intl]);
 
   return (
-    <Box style={{ height: '215px', width: '200px', overflow: 'hidden' }}>
+    <Box style={{ height: '230px', width: '200px', overflow: 'hidden' }}>
       {getStatusOptions}
     </Box>
   );
-});
+};